optimize config and remove rpaddle
