--- config/methods_config/rpaddle2.yaml
@@ -1,13 +1,0 @@
-METHOD:
-  name_method: 'RPADDLE2'
-  iter: 1000
-  lambd: 75
-  lr : 0.001 # 0.001 for adam, 0.1 for sgd
-  beta : 2.
-  kappa : 0
-  id_cov : True
-  optimizer: 'adam'
-  acc_clustering: False
-  tunable: False
-  plot: False
-  temp: 1
